+++ src/SMServer/shared/Constants.swift
@@ -73,7 +73,13 @@ struct Const {
 	]
 
 	static let api_config: String = "config"
-	static let api_match: String = "match"
+	static let api_match_keyword: String = "match"
+	static let api_match_type: String = "match_type"
+	
+	static let api_match_vals = [
+		api_match_keyword,
+		api_match_type
+	]
 
 	static let custom_css_path = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)[0].appendingPathComponent("smserver_custom.css")